requirements.txt
dash
python-dotenv
psycopg2-binary
boto3
pandas